v3 docs: Fix droplet field on build example JSON
In 09f10a403c6bf7142a92eefabb5d60f93a6fd68e, we moved the `href` on the
`build.droplet` field into the `links` section. This makes it consistent
with how we present similar links on all other resources. This change
updates the example JSON in the docs to reflect this.

[#173368325]

Authored-by: Reid Mitchell <rmitchell@pivotal.io>