Switch blueprint for Event to use organization_guid
Previously the blueprint for the `event` object exposed and defaulted
the `space` attribute. This is a vestige from when the `event` model was
originally created; at the time, an event was only associated with a
space. Since then, an event can be associated with an organization
and/or a space through the `space_guid` and `organization_guid`, but the
blueprint was never updated.

With this commit, `event` mocks will now default the `organization_guid`
if not provided, rather than `space`. This has the side effect that the
default event (ie `Event.make`) will now be scoped to an organization
rather than a space. Space-scoped events can be created by specifying
both the `space_guid` and `organization_guid` field.

Authored-by: Belinda Liu <bliu@pivotal.io>