V3: valid timestamps → subset of ISO 8601
Timestamps are typically used when querying audit_events, e.g.

 `GET /v3/audit_events?created_at=2020-06-30T23:49:04Z`

We enforce an opinionated subset of ISO 8601:

- must match `YYYY-MM-DDThh:mm:ssZ`
  - no subseconds
  - no local timezones

We updated our tests to accommodate the "no local timezones" stricture
(we used `Time.now.utc` to force UTC).

We did a simple regex to guard against bad timestamps. It is admittedly
ugly, but late is the hour.

[#173542711]

Authored-by: Brian Cunnie <bcunnie@vmware.com>