Raise new UniquenessError on duplicate app names
We recently did some explorations into how we can better use v3 API error
`title`s + `code`s to allow clients to distinguish between 422s, which are
returned in many cases, without needing to resort to the `detail`
field. The result was that we wanted to add new categories of 422
errors, identified by their `title`/`code`. These errors will still have
a `detail` field that contains their original, longer message.

The first category that we're introducing is the `UniquenessError` (code
`10016`). We now raise it when trying to create an app with a name that
already exists in the target space. Previously, this had title
`UnprocessableEntity` (code `10008`), which did not give any more useful
information than the 422 HTTP status code already gave. The error
`detail` is unchanged.

This commit also introduces some patterns and helpers that will
hopefully be useful as we do more of this kind of thing: we have
`api_error!` and `v3_api_error!` methods that can raise new `ApiError`s
and `V3::ApiErrors`, respectively. And we have corresponding
`raise_api_error` and `raise_v3_api_error` RSpec matchers to make it
easier to assert that we raised the right error.

[finishes #173506370]

Authored-by: Reid Mitchell <rmitchell@pivotal.io>