Make BackgroundJobEnvironment specs work on linux
Also improve the test for the case without a readiness port. The
previous test only tested that a single specific port was not opened,
when any port may have been opened. The new test ensures that the same
number of ports are open before and after running the method. This may
cause problems when running tests in parallel, but we can address that
problem when/if it occurs

Authored-by: Matt Royal <mroyal@pivotal.io>