Handle case where blobstore_key is not set on droplet
When downloading a droplet, if the blobstore key is not set, you would
previously get a BlobNotFound error. This can make it hard to
distinguish this case from the other cases where BlobNotFound errors are
raised. We want to make it clearer when, e.g. staging does not complete
and our droplet does not have a blobstore key.

Co-authored-by: Renee Chu <rchu@pivotal.io>
Co-authored-by: Merric De Launey <mdelauney@pivotal.io>
Co-authored-by: Reid Mitchell <rmitchell@pivotal.io>