Create a droplet in the kpack stager
- Previously for kpack apps, the droplet was created when the build
controller would send a patch request to CAPI after a successful OCI
Image build. We wanted to create the droplet when the Image CR is
created so that the Image CR would have a droplet guid label that could
be used by the image controller to know which droplet to update after a
stack update.
- Removed droplet creation from the BuildUpdate action
- Non-kpack lifecycle app staging workflows should not have been changed or
impacted by this change.

[172542372]

Co-authored-by: Piyali Banerjee <pbanerjee@pivotal.io>
Co-authored-by: Nat Bennett <nbennett@pivotal.io>