Dockerfile shouldn't know where configs will live
We added a new secrets file that is needed in practice on k8s. Rather
than hard-code the location of this file into the Dockerfile, we're just
going to remove all knowledge of where the config files live.

The k8s config determines where the config files are mounted, so it
should also specify their location

[#173973874]

Authored-by: Matt Royal <mroyal@pivotal.io>