v3(services): Fix job retrieval for service route binding
The job stores the resource_type in the db and uses it upon retrieval to
lookup in the correct db table. As resource type string and model name have
diverged now, we need to tell the Job model what the correct model is
otherwise we get a `table does not exist` error.

[#174398276](https://www.pivotaltracker.com/story/show/174398276)