add error on cf-for-k8s when type: buildpack is configured
this error is conditional on default_app_lifecycle for now as its the
most closely associated config field currently available. it might be
surprising to find that setting a default_app_lifecycle excludes the
opposite one.

in the future we should consider an explicit field about which lifecycles are
supported OR find a way to support the user asking for either when the
other is not present.

[finishes #174224146]

Authored-by: Connor Braa <cbraa@pivotal.io>