Remove the `diego` config when on k8s
We use the presence of the `KUBERNETES_SERVICE_HOST` environment
variable to detect if we're running on k8s. If we are, we don't want to
include the `diego` config key in our schema.

We started with another approach, where we looked for `/var/vcap` on
disk to determine if we were in a BOSH VM. Upon further reflection we
realized that not all VM-based deployments of CF are bosh-based, and
this would likely cause problems.

Our hope is that this environment variable based solution will avoid
this problem.

[#175080273]

Co-authored-by: Matt Royal <mroyal@pivotal.io>
Co-authored-by: Nat Bennett <nbennett@pivotal.io>