Update rubocop to 0.89.0
- Fairly big jump because of various bugs with rubocop between 0.83 and
0.89
- Disable Style/RedundantRegexpCharacterClass and
Style/RedundantRegexpEscape to avoid messing with existing regular
expressions
- Disable Lint/BinaryOperatorWithIdenticalOperands because it errors
- Disable Lint/EmptyConditionalBody because there are places where it is
useful to document possible cases with empty conditional bodies
- Disable Lint/MissingSuper. This might be nice to enable in the future,
but it involves too many changes for this commit.
- Disable Style/CaseLikeIf because I don't care about this style.
- Disable Style/OptionalBooleanParameter. This might be nice to change
in the future, but it involves changing to many method interfaces for
this commit.
- Disable Style/RedundantAssignment because "redundant" naming of return
values is useful for documentation
- Disable Style/StringConcatenation because it involves too many changes
for this commit. The auto correct also messes up nested quotes.
- Disable Bundler/OrderedGems because it clobbers gem upgrade history
for marginal gain.

Authored-by: Greg Cobb <gcobb@pivotal.io>