Update rubocop to 0.90.0
- Disable Lint/UselessMethodDefinition because adhering to it breaks things. I
believe this is mostly because we are overriding constructors to
suppress superclass behavior. This is probably not great design, but
fixing it is outside the scope of this commit.
- Disable Style/CombinableLoops because it requires some significant
restructuring. Also there is some async processing in
diego/reporters/instances_reporter.rb that might break if the loops are
combined.
- Disable Style/KeywordParametersOrder to avoid changing existing method
interfaces.
- Disable Style/SoleNestedConditional because it involves some
significant restructuring. This might be a good thing to enable in the
future.

Authored-by: Greg Cobb <gcobb@pivotal.io>