v3(services): space filtering include org offerings
There was an issue where filtering service offerings by space_guids
would not include org-restricted offerings, which was inconsistent with
service plans and with the design for V3 service visibility. This was
corrected in a previous commit by sharing code between the offering and
plan fetchers, and this commit adds integration testing to check the
behaviors for offering and plans, with space_guids and
organization_guids.

[#175415141](https://www.pivotaltracker.com/story/show/175415141)