Prevent files with insufficient permissions from being cached
- Previously, we were altering file permissions when we unzipped the
resources for an app to give ourselves read/write access. The altered
files were then stored in the resource cache
- When attempting to push the app containing a file with insufficient
permissions for a second time, a resource match would be found. We
preserve the mode provided in the fingerprint given to us from the
client, rather than using the altered mode the file was cached with.
(endpoint: POST /v3/resource_matches)
- When the client reached the upload bits portion of push, it would pass
along the fingerprints matched in the step above. The request would then
error out, as it would run into a check we have to ensure that the mode
on all fingerprints is >= 0600. (endpoint: POST /v3/packages/:guid/upload)

To solve this issue we have done the following:
- Moved the portion of code that altered permissions on unzipping the
application. This is still necessary in allowing CAPI to clean up the
unzipped files, and is now done as part of `remove_dirs_from_zip`. If
changing the permissions fails, an error is logged but it does not cause
the call to fail.
- Enforced a file permission check before uploading any files to the
resource cache. This prevents files with permissions < 0600 from being
uploaded in the first place, ultimately preventing a resource match from
ever being returned for insufficiently permissioned files.

Users who have previously uploaded an insufficiently permissioned file
may need to clear their resource cache as these changes do not account
for any files already in the cache.

Github issue: https://github.com/cloudfoundry/cloud_controller_ng/issues/1705
[finishes #173541786]

Authored-by: Sarah Weinstein <sweinstein@pivotal.io>