Staging no longer requires process types
- Staging will no longer fail if no process types are returned from the
buildpack as long as the latest web process already has a start command.
- The `with_start` argument is used in the v2 staging flow and bypassed
the requirement for process types, as long as the process already has a
start command.
- Some buildpacks (for example the python buildpack) do not return
process types unless a Procfile is provided. The previous logic meant
that staging would always fail without a Procfile in the v3 staging flow.
- Requiring a Procfile in the base case is an upgrade barrier for the v3
API. Instead, we can use the same default command as the v2 staging
flow. As a side effect, the process_types/start command logic no longer
depend on `with_start`
- Change the `app.processes.first.command.blank?` check to use
`app.newest_web_process` instead, since that is the process that will be
started (either later in `staging_completion_handler` in the `with_start`
flow or when assigning the droplet/starting the app in the v3 staging
flow).

[#1922]

Authored-by: Greg Cobb <gcobb@pivotal.io>