Scaling process memory no longer causes downtime
`/v3/processes/:guid/actions/scale`
`/v3/apps/:guid/processes/:type/actions/scale`

We set `skip_process_version_update` to true if we are calling this
endpoint to avoid creating a new version of the process. This
prevents the clock from propagating the desired process state to diego
as part of this command.

We chose to set the flag in the v3 controller to ensure this does not
change v2 behavior. We feel that after we expand this behavior to other
endppoint we may want to refactor out the use of this boolean since it's
essentialy adding a double negative.

[#175190738]

Co-authored-by: Belinda Liu <bliu@pivotal.io>
Co-authored-by: Merric de Launey <mdelauney@pivotal.io>