Validate JobPollableModel is Updated before DelayedJob Adding test to validate that the JobPollableModel insert happens before the DelayedJob insert when an asynchronous task is enqueued.
This is done by adding callbacks that wrap the DelayedJob before
Callback and also execute before the after Callback to extract the
current state of each model as the callback lifecycle progresses.
This test is designed to validate the elimination of the race
condition identified in issue #1973 and prevent reoccurrence.