Changing route destination ports should not cause downtime
If a user makes changes to the ports on a route, this will no longer
trigger a process version change and cause downtime when the runtime
makes updates. A restart is now required for new ports to be exposed on
the runtime level. This commit is part of a larger track of work around
zero-downtime for certain operations.

[#175190769](https://www.pivotaltracker.com/story/show/175190769)

Co-authored-by: Belinda Liu <bliu@pivotal.io>
Co-authored-by: Sarah Weinstein <sweinstein@pivotal.io>