services: allow longer service instance names (#2131)
Issue #2111 documents a problem with service instance names being
limited to 50 characters and a case to increase that limit.

The original reason for this limit was not documented when it was
introduced in 2013. The best working theory is that there was a service
broker used in Pivotal Web Services (PWS) that was not compatible with
names longer than 50 characters. That is no longer the case. But there
is a risk that this change could cause issues for some service brokers.
We note that OSBAPI does not specify a limit.

The limit has been increased to 255 characters so that it matches the
limit for service binding names.

[#176818363](https://www.pivotaltracker.com/story/show/176818363)