Add rake task to generate large amounts of service data
For assessing performance of service related endpoints it is useful to
be able to generate large amounts of data about services, plans,
instances etc.

Add a rake task to do this to replace the slow script that was piped
into bin/console as this suffered from being a TTY/slow loading of
commands. To speed up inserts, <Class>.new is used instead of
<Class>.create as this way we can multi_insert which uses a single DB
query to do the whole insert rather than 1 per object which gets slow
from so many round-trips to the DB.

Co-Authored-By: Florian Braun <FloThinksPi@users.noreply.github.com>