Ignore removal of processes values in manifest diffs
Previously if a memory of diffs field was removed from an manifest, the
behavior would be to display the removal in the manifest diff. However,
manifest diffs operate in an additive fashion, and the removal of the
values does not set the attributes to the default, but leaves them the
same.

These changes reflect the fact that these values are remaining unchanged
by way of the manifest diff.

[#177552249]

Co-authored-by: Marc Paquette <mpaquette@pivotal.io>
Co-authored-by: Sarah Weinstein <sweinstein@pivotal.io>
Co-authored-by: Teal Stannard <tstannard@pivotal.io>