Ignore diffs for processes that are not represented in the manifest
Previously, if a process was removed from the manifest our diffing logic
would error when trying to index into manifest_app_hash_process. As the
processes are not actually removed when applying a manifest or pushing
an application, and working on the assumption that end users would like to
see the impact that their manifest will have on the running state of
their app, we are omiting this diff.

[#177552249](https://www.pivotaltracker.com/story/show/177552249)

Authored-by: Sarah Weinstein <sweinstein@pivotal.io>