Don't eagerly load service plan visibilities on list endpoint
The only place that the visibilities relationship is used is in
determining what "visibility_type" a plan has and that only needs to
check `.empty?` on the visibilties. By not eagerly loading the
visibilities we save having to query and parse a huge amount of data
(num_visibilities ~= num_orgs * num_plans). When this data is needed to
check for `empty?` the ORM will instead issue a `SELECT 1 ... WHERE
service_plan_id = '1234'` for each plan. Even in the scenario where a
query is needed for every plan (i.e. the plan is neither public nor
space-scoped), this is far more efficient than fetching and parsing the
actual table contents. We have seen an improvement from 90s for
/v3/service_plans?per_page=500 to 1.3s, even in the worst-case scenario
of every plan having visibility_type of organization.

Co-Authored-By: Florian Braun <FloThinksPi@users.noreply.github.com>