Prevent large page/per_page numbers from throwing DB errors
Pagination options eventually become LIMIT/OFFSET queries in SQL. In
Postgres the OFFSET parameters is a bigint which has a maximum value
which can be exceeded by querying for a large page size (and/or a large
per_page). This limit is higher in MySQL/MariaDB but the number is large
enough that using Postgres as the lowest common denominator seems okay.
This should ensure that requests with very large page sizes receive a
400 and a useful error message rather than a 500 and no debugging
information (DB errors only show up in CC/DB logs)

Co-Authored-By: Florian Braun <FloThinksPi@users.noreply.github.com>