Filter security group spaces in Ruby not SQL
For users with access to a large number of spaces, the SQL query string
for checking which spaces a security group applies to gets quite big
(sometimes >1MB). By grabbing all the spaces a security group applies to
and doing the filtering in Ruby, this query string is much smaller. The
potential extra cost of running the `include?` seems to be offset by the
saving in creating/sending the large query string and using the
`select(:guid)` also saves some data transfer to further improve the
latency of this query.