Fix process and sidecar diffs
Use a custom similarity function that compares the value of the "type"
key (for processes) or "name" key (for sidecars) to determine if two
hashes are sufficiently similar (and should be considered a replace
instead of a remove/add).

Returns nil otherwise so that the diff logic will fall back to the
built-in similarity algorithm.

Fixes #2248

[#177919811]

Co-authored-by: Tom Viehman <tviehman@pivotal.io>