WIP: Swap Thin for Puma
Still needs a lot of work:
* Not feature flagged (probably should be)
* Periodic metrics are added on every worker but most are gauges about
the individual thread (see vitals) so will clash and cause confusing
numbers as each worker reports
* Starting EM in a different thread means it _may_ not get shutdown
properly but given it just does metrics now I'm not sure this really
matters anymore?
* Are we really benefitting from preload_app! - we don't really need
fast startup time for workers as we probably won't cycle them and it
means we have to disconnect the DB after loading the models. It might be
simpler to just start the CC after forking which should hopefully mean
the DB only connects after forking
* Some config options like DB connection pool will need
expressing/updating to be per-worker to avoid over-allocating