Add GUID as secondary order parameter when available
This change resolves an issue with inconsistent ordering that CAPI users
were experiencing when resources were created within the same second as
each other.  Resources that have a GUID field in the CAPI DB will now
order by GUID as a secondary order parameter, which should cause the
ordering of resources to be consistent.

https://github.com/cloudfoundry/cloud_controller_ng/issues/2249

Co-authored-by: Jenna Goldstrich <jgoldstrich@pivotal.io>
Co-authored-by: Sarah Weinstein <sweinstein@pivotal.io>