Allow application supporters to use route endpoints
The readable_org_guids_for_domains function takes a flag for now about
whether to include application supporters. All usages of this function
will eventually set this flag to true at which point it can be removed.

As some controllers start using the untrusted_* functions, the
user_helpers spec file needs to return the correct value for whichever
of these functions has been called. The tests that use this helper do
not need to know which of the two functions was called, only that the
permission will or will not be granted.

Added a has_application_supporter? public function to the space object
so that we can use untrusted_can_read_route? as the can_read_route?
function is used by v2 endpoints which we don't want to grant access to
for application supporters

Co-authored-by: Philipp Thun <philipp.thun@sap.com>