Remove perm project code
The perm project has been discontinued and in the cloudfoundry-attic for
a few years now so it is safe to say nobody is running it. As we work on
permissions, it simplifies things to not have to worry about replicating
behaviour to the perm permissions files and we can combine the queryer
and the cloud_controller/permissions files together

Process followed was:
* Remove perm/*.rb files
* Migrate cloud_controller/permissions.rb + queryer.rb ->
permissions_queryer.rb including spec tests
* Fix places where queryer or permissions object was constructed
* Fix up some places where perm was called explicitly (e.g.
tasks_controller) and remove redundant functions
* Fix tests
* Delete perm related gems
* Search for any other references to perm to check it was fully purged
(e.g. .travis.yml)