Log caller for SQL queries
Logs the calling function (of cloud_controller_ng code) for each DB
query if SQL logging is enabled. This should help track down which
functions are calling expensive queries and where excessive queries are
being made. Running some perf tests against this, it appears to have
little to no impact on the performance of CC since I suspect all this
information is readily available in the callstack and the regex isn't
too complex