Allow expiring of DB connections
Add some config parameters which (only when provided) enable the
connection_expiration extension for Sequel. This allows connections to
be closed and recreated after a certain time period (with some random
delay to prevent the entire pool being recycled at once).

This should help DBs that are suffering from performance issues related
to extremely long running connections as the recycling should allow
resources associated with those connections to be freed