Replace readable_org_guids_for_domains with query
Rather than creating the entire list of org guids in Ruby and passing
that as a giant filter to subsequent SQL queries, just create a subquery
that specifies which orgs are allowed and let the DB sort out the most
efficient way to handle it. This isn't really compatible with the perm
project but since that is deprecated I think it is fine to only
implement it for the db_permissions object.

We have seen extremely large query strings when an admin or a power-user
in many orgs requests domains as the entire list of (sometimes >100k)
guids is turned into a giant `WHERE IN ('abc', 'def' etc.)` SQL query.
This is expensive for Ruby to turn into the query, expensive to send
over the network and expensive for the DB to process.

Databases are very very good at working out the most efficient way to
run a query so even though the subqueries produced by this look pretty
complex (due to each role being a separate table and space roles needing
a join to the orgs table), the actual runtime is much better this way.