Log caller and query length for sql queries (#2329)
* Log caller for SQL queries

Logs the calling function (of cloud_controller_ng code) for each DB
query if SQL logging is enabled. This should help track down which
functions are calling expensive queries and where excessive queries are
being made. Running some perf tests against this, it appears to have
little to no impact on the performance of CC since I suspect all this
information is readily available in the callstack and the regex isn't
too complex

* Add extension to log SQL query length

We've seen some queries get extremely long (>1MB) which causes CC to run
slow whilst building the query. To help monitor this and fix endpoints
generating large queries, log the character length of the query in the
SQL query logging string

Co-authored-by: johha <45264872+johha@users.noreply.github.com>
Co-authored-by: andy-paine <andy.paine@engineerbetter.com>