Show when instances get rescheduled in `cf events`
This commit goes alongside work to TPS and runtimeschema. Together
they allow CF-for-VMs users to see when their app instances get
evacuated from Diego Cells.

As a user running a complex web service on top of Cloud Foundry, it
can be confusing when apps seem to restart for no reason. Even once
users understand this happens during Diego updates, they still have no
automatic insight into when that happened.

The audit event is called `audit.app.process.rescheduling` so that is
quite generic to whatever container scheduler is used behind the
scenes. For instance Eirini could talk to this endpoint and provide
a different `reason` string to explain the rescheduling.