Add connection expiration to deployment_updater_schema
The deployment updater uses the same DB setup classes as the main
application and is deployed as a long running process so may well
benefit from expiring connections. Add it to the schema so that it can
be configured correctly.

Also remove an unnecessary if since it is already inside an if block