Enforce service name uniqueness in shared services in spaces
Co-authored-by: Brandon Potts <bpotts@vmware.com>
Co-authored-by: Seth Boyles <sboyles@pivotal.io>
Co-authored-by: Jenna Goldstrich <jgoldstrich@pivotal.io>

This commit prevents a space from being shared if it contains a
service instance whose name is already in use in the target space.

This commit resolves GitHub issue #2306

This commit is a retry of a previous commit.

Co-authored-by: Brandon Potts <bpotts@vmware.com>
Co-authored-by: Seth Boyles <sboyles@pivotal.io>

Fix issue in unit tests where target_space would no longer update its
list of shared_services

Remove test that validated idempotant service_instance creates, this is
no longer valid as we are changing them to enforce name uniqueness

Authored-by: Jenna Goldstrich <jgoldstrich@pivotal.io>