Remove unnecessary select from users table
Every request to the CC goes through the SecurityContextConfigurer that
fetches the user from the database. But instead of a single

  SELECT * FROM "users"
    WHERE ("guid" = '...')

there is an additional

  SELECT * FROM "users"
    WHERE (("guid" = '...') AND ("is_oauth_client" IS TRUE))

For oauth clients the second where condition is

  "is_oauth_client" IS FALSE

With this small refactoring, the user is only fetched once and the
additional check (is_oauth_client == true / false) is done based on the
already fetched data. Although this is a very cheap select statement, it
is better not to send it to the database. This will speed up every call
(by a very small amount).