Do not treat errors from CloudController modules as UnknownErrors
Every error that was neither a compound error (responding to
'underlying_errors'), nor an api error (responding to 'code'), nor a
service error (type StructuredError) was treated as UnknownError by the
V3ErrorHasher. This means that for all other errors the not so helpful
message "An unknown error occurred." was returned to the client.

But as there are many explicitly defined StandardErrors spread
throughout the cloud controller code base, it makes more sense to
differentiate between these concrete (and known) errors and other e.g.
RuntimeErrors.

With this change all StandardErrors from modules 'CloudController' or
'VCAP::CloudController' are treated as known errors and presented in a
standard way:
 title: CF-<name of error class>
 detail: <error message>
 code: 10001

Additionally this change contains some renamings and refactorings for
consistency.