Update request counts in database only every n requests
To reduce lock contention on individual rows of the request_counts
database table, this commit introduces a new config option
rate_limiter.update_db_every_n_requests. If set to 1, rate limiting
behaves as before; but if set to a higher value - e.g. 10 - this means
that the database will only be updated every 10 requests per user and
API node. The returned X-RateLimit-Remaining header will show a value
that slightly varies depending on the API node hit by the user's
request, but the overall rate limit semantics remain the same.