Improve efficiency of pagination count query
When paginating results, a COUNT is performed to get the total number of
results that would be returned by the query. For queries with a
DISTINCT in their SELECT, this requires the DB to perform a full
sort over the entire SELECT table.* columns (sometimes 15+) which can be
very slow. Since this is only being used for a COUNT, only a unique
identifier is needed to do the count which is already unique so does not
require sorting. This halves the EXPLAIN cost of some of the
service_plans queries since that is currently sorting over 15+ fields
just to return a single COUNT.

Co-Authored-By: Philipp Thun <philipp.thun@sap.com>