Only join required tables
ServicePlanListFetcher and ServiceOfferingListFetcher JOIN multiple
tables to be used in WHERE conditions. The actual set of tables to be
joined depends on the provided filters. Instead of joining a superset of
tables, this change ensures that only the required tables will be
joined.

Unfortunately Postgres does not remove unnecessary JOINs and so they
possibly result in slower query execution times.

Before using another table in a WHERE clause, one simply has to call
join_<table>(dataset). This can even be done multiple times for each
table as the underlying join(dataset, type, table, on) function ensures
that only unique table expressions are added.