Use COUNT(*) OVER window function when paginating lists
This window function (supported by both MySQL > v8 and all modern
Postgres versions) allows fetching of the total number of results a
query would return (OVER the whole window), ignoring the LIMIT/OFFSET
clauses from the pagination. This means that even if the pagination only
returns 50 results, the `total` field will contain e.g. 120 if that's
how many the filter would normally return. This means only a single
SELECT is needed rather than a SELECT COUNT(*) + SELECT which should
massively lower the load on the DB (and make us less reliant on cache
hits in the DB). As the pagination extension is only doing the
limit/offset in this scenario we can do it ourselves and avoid overhead
of extending the dataset to include the plugin.

The eager_graph plugin uses joins to fetch related data in a query and
then maps the joined entities into associations on the main object. As
part of doing this, it expects to know what all the fields of the base
model are so that it can tell what to extract from each row into the
base model. For eager_graph loaded queries, this was resulting the in
the :pagination_total_results field being thrown away. To avoid this
issue, we register the :pagination_total_results field as part of the
main table using add_graph_aliases then remove it before returning the
final records.

Additionally, as part of making this change, the get_page function has
been updated to use some nice Sequel constructs like order_append and
qualify to simplify the code and avoid string interpolation.

Co-Authored-By: Philipp Thun <philipp.thun@sap.com>