Fix ordering of buildpacks
The sequel paginator uses the .order function which overrides any
ordering already applied to the dataset. For buildpacks, the :list
plugin means the Buildpack.dataset has a default ORDER BY of position.
To avoid this, make sure the pagination_options contain the correct
ordering.

This breaks the test for the Kpack buildpacks list endpoint but this
test has always been wrong as the returned results should be ordered by
the order specified in the builder, not ordered by :id (which was the
default in pagination_options for buildpacks before)

Co-Authored-By: Johannes Haass <johannes.haass@sap.com>