Use V3ErrorHasher in LoggingContextJob when originating in a v3 endpoint
The LoggingContextJob does not use the correct ErrorHasher based on the
endpoint (i.e. v2 vs v3) of the originating request. This change makes
the api version available to the LoggingContextJob (via Enqueuer) by
setting it in the middleware (similar to the VCAP-Request-ID). No
dedicated middleware was introduced; instead it has been added to the
existing VcapRequestId middleware as it can be removed when v2 is
discontinued.

Besides using the api version from the current thread in the
LoggingContextJob to decide whether V2ErrorHasher (default) or
V3ErrorHasher should be used, the same is done in other places where
this differentiation is needed (i.e. SecurityContextSetter,
RateLimiter).

When checking for a 'compound_error?' the BaseErrorHasher uses the
instance type rather than duck-typing as the V3ErrorHasher can only
handle CompoundErrors correctly, i.e. it expects 'underlying_errors' to
be ApiErrors with fields message, name and code.