Consistent error handling in background jobs
This change tackles inconsistent error handling in background jobs.

After adding a test to the enqueuer_spec.rb (AppDelete raises
SubResourceError), it could be seen that the DeleteAppErrorTranslatorJob
only effects the cf_api_error in the PollableJobModel (i.e. saved by
PollableJobWrapper), whereas the cf_api_error in Delayed::Job contained
'UnknownError'.

In a different error scenario (AppDelete raises BlobstoreError), the
error was 'translated' inside the LoggingContextJob. As a result, it was
stored consistently in both PollableJobModel and Delayed::Job.

With this change the DeleteAppErrorTranslatorJob was removed and 'error
translation' for background jobs is solely done in LoggingContextJob.

The V3ErrorsHelper mixin was renamed and moved to a dedicated file.
Additional convenience methods were added (e.g. blobstore_error and
compund_error).

In application_controller_spec.rb some methods needed renaming to
prevent conflicts with the ErrorsHelper mixin.