Rate limit endpoints that interact with service brokers
Limit the maximum number of concurrent requests per-user to a VM that
may interact with a service broker. This can help mitigate the webserver
request queue filling up with requests waiting on slow service brokers.

Rather than sending rate limit headers as the main API rate limiter
does, instead this just returns a nicely formatted error (based on v2 or
v3 formatting) as it is concerned with concurrency rather than max
limits over time with resets etc.

Some requests will hit these endpoints (e.g. v3 user provided service
instances) that don't require access to a broker but these should be so
fast that the concurrency limit is not as much of a problem.

0 or negative configuration for the max value is taken to mean "do not
enable this feature"

Signed-off-by: Andy Paine <andrew.paine@sap.com>