Use an optimised conversion proc for Postgres timestamps
When looking at profiling data one can see that the timestamp conversion
is mainly taking place in Sequel::SequelMethods::convert_input_timestamp
and that the string returned from the database is actually parsed twice.
The second parsing is done to check if the given string contains an
offset - which is never the case as Postgres field type "timestamp
without time zone" is used. So what happens boils down to something like
this:

  t = Time.parse(value)
  Date._parse(value).has_key?(:offset)
  (t + t.utc_offset).utc

Benchmarking this locally (1 million times) shows a (reference) time of
21.6 seconds.

Removing the unnecessary second parsing reduces the time to 12.7s.

As the timestamp format is fix (i.e. no heuristics are needed), the
parsing can be done with strptime instead of parse. To distinguish
between timestamp strings with and without fractional seconds, it is
checked whether they contain a '.':

  DateTime.strptime(
    value,
    value.include?('.') ? '%Y-%m-%d %H:%M:%S.%N' : '%Y-%m-%d %H:%M:%S'
  ).to_time.utc

The measured benchmark time for this conversion is 4.0s.

By using DateTime._strptime to parse the timestamp and Time.utc() to
construct a new Time object, the time can be reduced to 2.8s (for 1
million conversions, i.e. a factor of ~8):

  dt = DateTime._strptime(
    value,
    value.include?('.') ? '%Y-%m-%d %H:%M:%S.%N' : '%Y-%m-%d %H:%M:%S'
  )
  Time.utc(
    dt[:year], dt[:mon], dt[:mday],
    dt[:hour], dt[:min], dt[:sec] + (dt[:sec_fraction] || 0)
  )

With this change the conversion proc is overwritten by Sequel::Postgres
::DatabaseMethods.add_conversion_proc.