Add sequel_pg gem
From the README [1]: "sequel_pg overwrites the inner loop of the Sequel
postgres adapter row fetching code with a C version. The C version is
significantly faster than the pure ruby version that Sequel uses by
default."

A local benchmark test (selecting a single user from the database, 10000
times) shows the following improvements:

- without sequel_pg:

  user       system     total
-------------------------------
11.120050  18.784816  29.904866
11.189934  18.721485  29.911419

- with sequel_pg:

  user       system     total
-------------------------------
 0.596703   0.256924   0.853627
 0.591460   0.250780   0.842240

According to the "PostgreSQL-specific Support in Sequel" documentation
[2] "sequel_pg [..] optimizes the fetching of rows, generally resulting
in a ~2x speedup. It is highly recommended to install sequel_pg if you
are using the postgres adapter with pg."

[1] https://github.com/jeremyevans/sequel_pg
[2] https://github.com/jeremyevans/sequel/blob/master/doc/
      postgresql.rdoc#sequel_pg-postgrespg-only-